+++ game/client/icliententityinternal.h
@@ -25,7 +25,7 @@ class ClientClass;
 //-----------------------------------------------------------------------------
 
 typedef CBaseHandle ClientEntityHandle_t;
-typedef unsigned short SpatialPartitionHandle_t;
+typedef int SpatialPartitionHandle_t;
